+++ src/components/basicComponents/GoBack.vue
@@ -1,13 +1,13 @@
 <template>
   <div class="go-back__container">
-   <ArrowLeft />
+    <ArrowLeft />
     <router-link to="/feedbacks">Go Back</router-link>
   </div>
   <slot></slot>
 </template>
 
 <script setup>
-import ArrowLeft from '../icons/ArrowLeft.vue';
+import ArrowLeft from "../icons/ArrowLeft.vue";
 // export default {
 //   props: {
 //     to: {
