+++ src/components/basicComponents/RoadmapMenu.vue
@@ -22,6 +22,6 @@
 </template>
 
 <script setup>
-import BaseBox from "../../BaseBox.vue";
+import BaseBox from "./BaseBox.vue";
 //idea for future solution: check the index of separate status and make a method for calculating the number of them if "planned" i++ etc.
 </script>
