+++ src/components/basicComponents/SortByCategory.vue
@@ -12,6 +12,6 @@
 </template>
 
 <script setup>
-import BaseBox from "../../BaseBox.vue";
+import BaseBox from "./BaseBox.vue";
 import SortByCategoryItem from "./SortByCategoryItem.vue";
 </script>
