+++ src/components/basicComponents/SortByCategoryItem.vue
@@ -1,0 +1,9 @@
+<template>
+  <li>
+    <BaseButton>Test</BaseButton>
+  </li>
+</template>
+
+<script setup>
+import BaseButton from "./BaseButton.vue";
+</script>
