--- src/components/ui/BlueBar.vue
@@ -1,12 +1,0 @@
-<template>
-  <div class="blue-bar__card">
-    <slot></slot>
-  </div>
-</template>
-
-<style scoped lang="scss">
-.blue-bar__card {
-  background-color: #373f68;
-  // bug to fix - there should be a global variable, I need to check the configuration of scss, if vite accept a new way of files organisation in scss (@forward and @use) - for this moment my configuration doesn't work so I am going to finish the sceleton and I'll check this problem later.
-}
-</style>
