--- src/components/ui/TheHeaderMenu/sortByCategory/SortByCategoryItem.vue
@@ -1,12 +1,0 @@
-<template>
-  <li>
-    <BaseButton>Test</BaseButton>
-  </li>
-</template>
-
-<script setup>
-import BaseButton from "../../BaseButton.vue";
-  // props: {
-  //   categories: ["All", "UI", "UX", "Enhancement", "Feature", "Bug"],
-  // },
-</script>
